+++ analysis/shamanrestoration/src/CombatLogParser.tsx
@@ -1,4 +1,3 @@
-import { Panel } from 'interface';
 import CoreCombatLogParser from 'parser/core/CombatLogParser';
 import ManaTracker from 'parser/core/healingEfficiency/ManaTracker';
 import LowHealthHealing from 'parser/shared/modules/features/LowHealthHealing';
@@ -15,7 +14,6 @@ import {
 } from '@wowanalyzer/shaman';
 
 import { ABILITIES_AFFECTED_BY_HEALING_INCREASES } from './constants';
-import Feeding from './Feeding';
 import Abilities from './modules/Abilities';
 import HealingDone from './modules/core/HealingDone';
 import HealingEfficiencyDetails from './modules/core/HealingEfficiencyDetails';
@@ -28,7 +26,6 @@ import Checklist from './modules/features/checklist/Module';
 import CooldownThroughputTracker from './modules/features/CooldownThroughputTracker';
 import MasteryEffectiveness from './modules/features/MasteryEffectiveness';
 import SpellUsable from './modules/features/SpellUsable';
-import StatValues from './modules/features/StatValues';
 import TidalWaves from './modules/features/TidalWaves';
 // Talents
 import EmbraceOfEarth from './modules/shadowlands/conduits/EmbraceOfEarth';
@@ -98,7 +95,6 @@ class CombatLogParser extends CoreCombatLogParser {
     castBehavior: CastBehavior,
     checklist: Checklist,
     spellUsable: SpellUsable,
-    statValues: StatValues,
 
     // Talents
     torrent: Torrent,
@@ -156,25 +152,6 @@ class CombatLogParser extends CoreCombatLogParser {
     chainHarvest: ChainHarvest,
     primordialWave: PrimordialWave,
   };
-
-  generateResults(options: any) {
-    const results = super.generateResults(options);
-
-    results.tabs = [
-      ...results.tabs,
-      {
-        title: 'Feeding',
-        url: 'feeding',
-        render: () => (
-          <Panel style={{ padding: 0 }}>
-            <Feeding cooldownThroughputTracker={this.getModule(CooldownThroughputTracker)} />
-          </Panel>
-        ),
-      },
-    ];
-
-    return results;
-  }
 }
 
 export default CombatLogParser;
