+++ analysis/shamanrestoration/src/constants.ts
@@ -106,8 +106,18 @@ export const ABILITIES_AFFECTED_BY_MASTERY = [
 ];
 
 export const ABILITIES_FEEDING_INTO_CBT = [
-  ...SHAMAN_BASE_ABILITIES,
-  SPELLS.ASCENDANCE_HEAL,
+  SPELLS.HEALING_WAVE,
+  SPELLS.CHAIN_HEAL,
+  SPELLS.HEALING_SURGE,
+  SPELLS.RIPTIDE,
+  SPELLS.HEALING_RAIN_HEAL,
+  SPELLS.WELLSPRING_HEAL,
+  SPELLS.UNLEASH_LIFE_TALENT,
+  SPELLS.EARTH_SHIELD_HEAL,
+  SPELLS.DOWNPOUR_TALENT,
+  SPELLS.ASCENDANCE_INITIAL_HEAL,
+  SPELLS.CHAIN_HARVEST_HEAL,
+  SPELLS.PRIMORDIAL_WAVE_HEAL,
   SPELLS.NATURES_GUARDIAN_HEAL,
 ];
