+++ analysis/shamanrestoration/src/modules/features/CooldownThroughputTracker.ts
@@ -1,0 +1,28 @@
+import SPELLS from 'common/SPELLS';
+import CoreCooldownThroughputTracker, {
+  BUILT_IN_SUMMARY_TYPES,
+} from 'parser/shared/modules/CooldownThroughputTracker';
+
+class CooldownThroughputTracker extends CoreCooldownThroughputTracker {
+  static cooldownSpells = [
+    ...CoreCooldownThroughputTracker.cooldownSpells,
+    {
+      spell: SPELLS.ASCENDANCE_TALENT_RESTORATION.id,
+      summary: [
+        BUILT_IN_SUMMARY_TYPES.HEALING,
+        BUILT_IN_SUMMARY_TYPES.OVERHEALING,
+        BUILT_IN_SUMMARY_TYPES.MANA,
+      ],
+    },
+    {
+      spell: SPELLS.CLOUDBURST_TOTEM_TALENT.id,
+      summary: [
+        BUILT_IN_SUMMARY_TYPES.HEALING,
+        BUILT_IN_SUMMARY_TYPES.OVERHEALING,
+        BUILT_IN_SUMMARY_TYPES.MANA,
+      ],
+    },
+  ];
+}
+
+export default CooldownThroughputTracker;
