+++ analysis/shamanrestoration/src/modules/talents/CloudburstTotem.tsx
@@ -90,15 +90,10 @@ class CloudburstTotem extends Analyzer {
   }
 
   subStatistic() {
-    const feeding = this.cooldownThroughputTracker.getIndirectHealing(
-      SPELLS.CLOUDBURST_TOTEM_HEAL.id,
-    );
     return (
       <StatisticListBoxItem
         title={<SpellLink id={SPELLS.CLOUDBURST_TOTEM_TALENT.id} />}
-        value={`${formatPercentage(
-          this.owner.getPercentageOfTotalHealingDone(this.healing + feeding),
-        )} %`}
+        value={`${formatPercentage(this.owner.getPercentageOfTotalHealingDone(this.healing))} %`}
       />
     );
   }
