+++ analysis/shamanrestoration/src/modules/talents/SurgeOfEarth.tsx
@@ -38,17 +38,11 @@ class SurgeOfEarth extends Analyzer {
     this.healing += event.amount + (event.absorbed || 0);
   }
 
-  get feeding() {
-    return this.cooldownThroughputTracker.getIndirectHealing(SPELLS.SURGE_OF_EARTH_HEAL.id);
-  }
-
   subStatistic() {
     return (
       <StatisticListBoxItem
         title={<SpellLink id={SPELLS.SURGE_OF_EARTH_TALENT.id} />}
-        value={`${formatPercentage(
-          this.owner.getPercentageOfTotalHealingDone(this.healing + this.feeding),
-        )} %`}
+        value={`${formatPercentage(this.owner.getPercentageOfTotalHealingDone(this.healing))} %`}
       />
     );
   }
